requirements.txt
asgiref
attrs
backports.zoneinfo
build
CacheControl
certifi
cffi
charset-normalizer
cleo
crashtest
cryptography
defusedxml
diff-match-patch
distlib
dj-database-url
Django
django-environ
django-filter
django-health-check
django-import-export
django-jazzmin
djangorestframework
dulwich
et-xmlfile
filelock
gscholar
gunicorn
idna
importlib-metadata
importlib-resources
installer
jaraco.classes
jeepney
jsonschema
keyring
Markdown
MarkupPy
more-itertools
msgpack
odfpy
openpyxl
packaging
pexpect
Pillow
pkginfo
pkgutil_resolve_name
platformdirs
poetry
poetry-core
poetry-plugin-export
psutil
psycopg2-binary
ptyprocess
pycparser
pyproject_hooks
pyrsistent
pyscopg2
pytz
PyYAML
rapidfuzz
requests
requests-toolbelt
SecretStorage
shellingham
sqlparse
tablib
tomli
tomlkit
trove-classifiers
typing_extensions
urllib3
virtualenv
whitenoise
xlrd
xlwt
zipp
